requirements.txt
aiohttp
certifi
chardet
icalendar
idna
python-dateutil
pytz
requests
six
urllib3
